% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_factor_levels}
\alias{map_factor_levels}
\title{Map factor level codes to their labels}
\usage{
map_factor_levels(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
A data frame with the integer level codes (\code{levels}) and their
string labels (\code{labels}).
}
\description{
Map factor level codes to their labels
}
\examples{
map_factor_levels(factor(c("low", "high", "low")))
}
