% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisecanceling-package.R
\docType{package}
\name{noisecanceling-package}
\alias{noisecanceling}
\alias{noisecanceling-package}
\title{noisecanceling: Psychometric analysis of likert scales: methods for controlling for acquiescence bias}
\description{
This package implements classical and modern (IRT) psychometric analyses that control for acquiescence bias. The package wraps a number of functions of the Psych package in order to facilitate the analysis of multidimensional scales .
}
\author{
\strong{Maintainer}: Primi Ricardo \email{rprimi@mac.com} (\href{https://orcid.org/0000-0003-4227-6745}{ORCID})

}
\keyword{internal}
